feat(screens): mejoras pantallas inicio,info, contacto y dueños de alojamiento terminado.
